wms merging conflicts i do not remember.
